src/containers/worldbank-container/requirements.txt
azure-storage-blob
requests
pandas
wbdata
country-converter
pyarrow